RavenDB-23085 - Started to implement shared jounrals - We can now submit a journal write from a branch env to a root env, which will record it in its own journal

+++ src/Raven.Client/Documents/Indexes/StorageCommitDetails.cs
@@ -2,8 +2,8 @@
 {
     public sealed class StorageCommitDetails 
     {
-        public int NumberOfModifiedPages { get; set; }
+        public long NumberOfModifiedPages { get; set; }
 
-        public int NumberOf4KbsWrittenToDisk { get; set; }
+        public long NumberOf4KbsWrittenToDisk { get; set; }
     }
 }
